requirements.txt
fastapi
uvicorn
pydantic==1.10.13
openai
pinecone
python-docx
pypdf
langchain-text-splitters
requests
tiktoken
python-dotenv  # To load variables from the .env file